requirements.txt
requests==2.31.0
pushbullet.py==0.9.1
